Discord dropdown concept

+++ api/internal/models/submission.go
@@ -9,7 +9,7 @@ type Submission struct {
 	DiscordID    string    `json:"discord_id"`
 	Party        []string  `json:"party"`
 	ActivityID   string    `json:"activity_id"`
-	PersonalBest Numeric   `json:"personal_best"`
+	PersonalBest int   	   `json:"personal_best"`
 	Approved     bool      `json:"approved"`
 	ImgurUrl     string    `json:"imgur_url"`
 }
